refactor: use BaseMenuItem in ActionListItem
