+++ packages/blade/src/components/BaseMenu/index.ts
@@ -1,0 +1,1 @@
+export { BaseMenuItem } from './BaseMenuItem/BaseMenuItem';
