+++ packages/blade/src/components/BaseMenu/types.ts
@@ -1,0 +1,37 @@
+import type { LinkProps } from '~components/Link';
+import type { BladeCommonEvents } from '~components/types';
+import type { Platform } from '~utils';
+
+type BaseMenuItemProps = {
+  as?: 'a' | 'button';
+  id?: string;
+  title: string;
+  description?: string;
+  className?: string;
+  isDisabled?: boolean;
+  isSelected?: boolean;
+  isVisible?: boolean;
+  isKeydownPressed?: boolean;
+  leading?: React.ReactNode;
+  trailing?: React.ReactNode;
+  titleSuffix?: React.ReactNode;
+  selectionType?: 'single' | 'multiple';
+  color?: 'negative';
+  href?: LinkProps['href'];
+  target?: LinkProps['target'];
+  rel?: LinkProps['rel'];
+  role?: 'menuitem' | 'menuitemcheckbox' | 'option' | 'link';
+  tabIndex?: number;
+  onClick?: Platform.Select<{
+    web: React.MouseEventHandler;
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    native: undefined | ((event: any) => void);
+  }>;
+} & BladeCommonEvents;
+
+type StyledBaseMenuItemContainerProps = Pick<
+  BaseMenuItemProps,
+  'color' | 'selectionType' | 'isSelected' | 'isDisabled' | 'isVisible' | 'isKeydownPressed'
+>;
+
+export type { BaseMenuItemProps, StyledBaseMenuItemContainerProps };
