+++ packages/blade/src/components/DatePicker/QuickSelection/QuickSelectionItem.web.tsx
@@ -1,11 +1,8 @@
-import { StyledActionListItem } from '~components/ActionList/styles/StyledActionListItem';
-import { Text } from '~components/Typography';
+import { BaseMenuItem } from '~components/BaseMenu';
 import { isReactNative } from '~utils';
-import { makeAccessible } from '~utils/makeAccessible';
-import { useIsMobile } from '~utils/useIsMobile';
 
 type QuickSelectionItemProps = {
-  children: React.ReactNode;
+  children: string;
   isSelected: boolean;
   onClick?: React.MouseEventHandler<HTMLButtonElement>;
 };
@@ -14,27 +11,15 @@ const QuickSelectionItem = ({
   onClick,
   isSelected,
 }: QuickSelectionItemProps): React.ReactElement => {
-  const isMobile = useIsMobile();
   return (
-    <StyledActionListItem
+    <BaseMenuItem
       as={!isReactNative() ? 'button' : undefined}
       selectionType="single"
-      hasDescription={true}
-      isMobile={isMobile}
-      isKeydownPressed={false}
       isSelected={isSelected}
       onClick={onClick}
-      isVisible
-      {...makeAccessible({
-        selected: isSelected,
-        current: true,
-        role: 'option',
-      })}
-    >
-      <Text size="small" weight="medium" color="interactive.text.gray.normal">
-        {children}
-      </Text>
-    </StyledActionListItem>
+      title={children}
+      role="option"
+    />
   );
 };
