+++ packages/blade/src/components/types.ts
@@ -16,6 +16,10 @@ type BladeCommonEvents = {
     web: React.MouseEventHandler;
     native: undefined | ((event: any) => void);
   }>;
+  onMouseUp?: Platform.Select<{
+    web: React.MouseEventHandler;
+    native: undefined | ((event: any) => void);
+  }>;
   onPointerDown?: Platform.Select<{ web: React.PointerEventHandler; native: undefined }>;
   onPointerEnter?: Platform.Select<{ web: React.PointerEventHandler; native: undefined }>;
   onTouchStart?: Platform.Select<{
